Crate eccodes_sys

Low-level Rust bindings for ecCodes

This is a -sys crate with raw, unsafe bindings to the library and its API should not be used directly. See the eccodes crate for high-level, safe bindings.

Due to the complexity of ecCodes library the decision has been made that this crate will not build ecCodes from source. See sections below for additional information how to install ecCodes on your system.

ecCodes is an open-source library for reading and writing GRIB and BUFR files developed by European Centre for Medium-Range Weather Forecasts.

Usage

This crate will look for existing libeccodes installation using pkg-config. The ecCodes library is then linked and bindings are generated using bindgen. If the library is not found, the build will fail.

ecCodes installation

The reccomended way to install ecCodes on your computer is using your package manager. For example, on Ubuntu you can use apt-get:

$ sudo apt-get install libeccodes-dev

Alternatively, you can install the library manually from source in suitable directory following this instructions.

Then add the lib/pkgconfig directory from your ecCodes installation directory to the PKG_CONFIG_PATH environmental variable. For example:

$ export PKG_CONFIG_PATH=<your_eccodes_path>/lib/pkgconfig

Features

There are two development features available:

docs - for documentation building, does not link ecCodes and includes bindings-docs.rs into lib.rs

tests - turns on generation of layout tests by bindgen, should not be used in production. Layout tests are off by default as they derefrence null pointers causing undefined behaviour

Structs

_G_fpos64_t

_G_fpos_t

_IO_FILE

_IO_codecvt

_IO_marker

_IO_wide_data

__fsid_t

__locale_data

__locale_struct

__mbstate_t

__pthread_cond_s

__pthread_cond_s__bindgen_ty_1__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2__bindgen_ty_1

__pthread_internal_list

__pthread_internal_slist

__pthread_mutex_s

__pthread_rwlock_arch_t

__sigset_t

__va_list_tag

bufr_descriptor

bufr_descriptors_array

bufr_descriptors_map_list

bufr_keys_iterator

codes_bufr_header

div_t

drand48_data

fd_set

grib_box

grib_context

grib_darray

grib_fieldset

grib_handle

grib_iarray

grib_index

grib_iterator

grib_key_value_list

grib_keys_iterator

grib_multi_handle

grib_nearest

grib_oarray

grib_order_by

grib_points

grib_sarray

grib_string_list

grib_util_grid_spec

grib_util_grid_spec2

grib_util_packing_spec

grib_values

grib_vdarray

grib_viarray

grib_vsarray

grib_where

itimerval

ldiv_t

lldiv_t

random_data

timespec

timeval

timezone

Constants

BIG_ENDIAN

BUFSIZ

BYTE_ORDER

CODES_7777_NOT_FOUND

CODES_ARRAY_TOO_SMALL

CODES_ATTRIBUTE_CLASH

CODES_ATTRIBUTE_NOT_FOUND

CODES_BUFFER_TOO_SMALL

CODES_CODE_NOT_FOUND_IN_TABLE

CODES_CONCEPT_NO_MATCH

CODES_CONSTANT_FIELD

CODES_CORRUPTED_INDEX

CODES_DECODING_ERROR

CODES_DIFFERENT_EDITION

CODES_DUMP_FLAG_ALIASES

CODES_DUMP_FLAG_ALL_ATTRIBUTES

CODES_DUMP_FLAG_ALL_DATA

CODES_DUMP_FLAG_CODED

CODES_DUMP_FLAG_DUMP_OK

CODES_DUMP_FLAG_HEXADECIMAL

CODES_DUMP_FLAG_NO_DATA

CODES_DUMP_FLAG_OCTECT

CODES_DUMP_FLAG_READ_ONLY

CODES_DUMP_FLAG_TYPE

CODES_DUMP_FLAG_VALUES

CODES_ENCODING_ERROR

CODES_END

CODES_END_OF_FILE

CODES_END_OF_INDEX

CODES_FILE_NOT_FOUND

CODES_FUNCTIONALITY_NOT_ENABLED

CODES_GEOCALCULUS_PROBLEM

CODES_HASH_ARRAY_NO_MATCH

CODES_INTERNAL_ARRAY_TOO_SMALL

CODES_INTERNAL_ERROR

CODES_INVALID_ARGUMENT

CODES_INVALID_BPV

CODES_INVALID_FILE

CODES_INVALID_GRIB

CODES_INVALID_INDEX

CODES_INVALID_ITERATOR

CODES_INVALID_KEYS_ITERATOR

CODES_INVALID_KEY_VALUE

CODES_INVALID_MESSAGE

CODES_INVALID_NEAREST

CODES_INVALID_ORDERBY

CODES_INVALID_SECTION_NUMBER

CODES_INVALID_TYPE

CODES_IO_PROBLEM

CODES_KEYS_ITERATOR_ALL_KEYS

CODES_KEYS_ITERATOR_DUMP_ONLY

CODES_KEYS_ITERATOR_SKIP_CODED

CODES_KEYS_ITERATOR_SKIP_COMPUTED

CODES_KEYS_ITERATOR_SKIP_DUPLICATES

CODES_KEYS_ITERATOR_SKIP_EDITION_SPECIFIC

CODES_KEYS_ITERATOR_SKIP_FUNCTION

CODES_KEYS_ITERATOR_SKIP_OPTIONAL

CODES_KEYS_ITERATOR_SKIP_READ_ONLY

CODES_LOG_DEBUG

CODES_LOG_ERROR

CODES_LOG_FATAL

CODES_LOG_INFO

CODES_LOG_WARNING

CODES_MESSAGE_MALFORMED

CODES_MESSAGE_TOO_LARGE

CODES_MISSING_BUFR_ENTRY

CODES_MISSING_DOUBLE

CODES_MISSING_KEY

CODES_MISSING_LONG

CODES_NEAREST_SAME_DATA

CODES_NEAREST_SAME_GRID

CODES_NEAREST_SAME_POINT

CODES_NOT_FOUND

CODES_NOT_IMPLEMENTED

CODES_NO_DEFINITIONS

CODES_NO_MORE_IN_SET

CODES_NO_VALUES

CODES_NULL_HANDLE

CODES_NULL_INDEX

CODES_NULL_POINTER

CODES_OUT_OF_AREA

CODES_OUT_OF_MEMORY

CODES_OUT_OF_RANGE

CODES_PREMATURE_END_OF_FILE

CODES_READ_ONLY

CODES_SECTION_BITMAP

CODES_SECTION_DATA

CODES_SECTION_GRID

CODES_SECTION_LOCAL

CODES_SECTION_PRODUCT

CODES_STRING_TOO_SMALL

CODES_SUCCESS

CODES_SWITCH_NO_MATCH

CODES_TOO_MANY_ATTRIBUTES

CODES_TYPE_BYTES

CODES_TYPE_DOUBLE

CODES_TYPE_LABEL

CODES_TYPE_LONG

CODES_TYPE_MISSING

CODES_TYPE_SECTION

CODES_TYPE_STRING

CODES_TYPE_UNDEFINED

CODES_UNDERFLOW

CODES_UNSUPPORTED_EDITION

CODES_UTIL_ACCURACY_SAME_BITS_PER_VALUES_AS_INPUT

CODES_UTIL_ACCURACY_SAME_DECIMAL_SCALE_FACTOR_AS_INPUT

CODES_UTIL_ACCURACY_USE_PROVIDED_BITS_PER_VALUES

CODES_UTIL_ACCURACY_USE_PROVIDED_DECIMAL_SCALE_FACTOR

CODES_UTIL_GRID_SPEC_LAMBERT_AZIMUTHAL_EQUAL_AREA

CODES_UTIL_GRID_SPEC_LAMBERT_CONFORMAL

CODES_UTIL_GRID_SPEC_POLAR_STEREOGRAPHIC

CODES_UTIL_GRID_SPEC_REDUCED_GG

CODES_UTIL_GRID_SPEC_REDUCED_LL

CODES_UTIL_GRID_SPEC_REGULAR_GG

CODES_UTIL_GRID_SPEC_REGULAR_LL

CODES_UTIL_GRID_SPEC_ROTATED_GG

CODES_UTIL_GRID_SPEC_ROTATED_LL

CODES_UTIL_GRID_SPEC_SH

CODES_UTIL_GRID_SPEC_UNSTRUCTURED

CODES_UTIL_PACKING_SAME_AS_INPUT

CODES_UTIL_PACKING_TYPE_GRID_COMPLEX

CODES_UTIL_PACKING_TYPE_GRID_SECOND_ORDER

CODES_UTIL_PACKING_TYPE_GRID_SIMPLE

CODES_UTIL_PACKING_TYPE_GRID_SIMPLE_MATRIX

CODES_UTIL_PACKING_TYPE_JPEG

CODES_UTIL_PACKING_TYPE_SPECTRAL_COMPLEX

CODES_UTIL_PACKING_TYPE_SPECTRAL_SIMPLE

CODES_UTIL_PACKING_USE_PROVIDED

CODES_UTIL_SET_SPEC_FLAGS_ONLY_PACKING

CODES_VALUE_CANNOT_BE_MISSING

CODES_VALUE_DIFFERENT

CODES_VERSION

CODES_WRONG_ARRAY_SIZE

CODES_WRONG_BITMAP_SIZE

CODES_WRONG_CONVERSION

CODES_WRONG_GRID

CODES_WRONG_LENGTH

CODES_WRONG_STEP

CODES_WRONG_STEP_UNIT

CODES_WRONG_TYPE

ECCODES_MAJOR_VERSION

ECCODES_MINOR_VERSION

ECCODES_REVISION_VERSION

ECCODES_VERSION

ECCODES_VERSION_STR

EOF

EXIT_FAILURE

EXIT_SUCCESS

FD_SETSIZE

FILENAME_MAX

FOPEN_MAX

FP_ILOGB0

FP_ILOGBNAN

FP_INFINITE

FP_NAN

FP_NORMAL

FP_SUBNORMAL

FP_ZERO

GRIB_7777_NOT_FOUND

GRIB_ARRAY_TOO_SMALL

GRIB_ATTRIBUTE_CLASH

GRIB_ATTRIBUTE_NOT_FOUND

GRIB_BUFFER_TOO_SMALL

GRIB_CODE_NOT_FOUND_IN_TABLE

GRIB_CONCEPT_NO_MATCH

GRIB_CONSTANT_FIELD

GRIB_CORRUPTED_INDEX

GRIB_DECODING_ERROR

GRIB_DIFFERENT_EDITION

GRIB_DUMP_FLAG_ALIASES

GRIB_DUMP_FLAG_ALL_ATTRIBUTES

GRIB_DUMP_FLAG_ALL_DATA

GRIB_DUMP_FLAG_CODED

GRIB_DUMP_FLAG_DUMP_OK

GRIB_DUMP_FLAG_HEXADECIMAL

GRIB_DUMP_FLAG_NO_DATA

GRIB_DUMP_FLAG_OCTECT

GRIB_DUMP_FLAG_READ_ONLY

GRIB_DUMP_FLAG_TYPE

GRIB_DUMP_FLAG_VALUES

GRIB_ENCODING_ERROR

GRIB_END

GRIB_END_OF_FILE

GRIB_END_OF_INDEX

GRIB_FILE_NOT_FOUND

GRIB_FUNCTIONALITY_NOT_ENABLED

GRIB_GEOCALCULUS_PROBLEM

GRIB_HASH_ARRAY_NO_MATCH

GRIB_INTERNAL_ARRAY_TOO_SMALL

GRIB_INTERNAL_ERROR

GRIB_INVALID_ARGUMENT

GRIB_INVALID_BPV

GRIB_INVALID_FILE

GRIB_INVALID_GRIB

GRIB_INVALID_INDEX

GRIB_INVALID_ITERATOR

GRIB_INVALID_KEYS_ITERATOR

GRIB_INVALID_KEY_VALUE

GRIB_INVALID_MESSAGE

GRIB_INVALID_NEAREST

GRIB_INVALID_ORDERBY

GRIB_INVALID_SECTION_NUMBER

GRIB_INVALID_TYPE

GRIB_IO_PROBLEM

GRIB_KEYS_ITERATOR_ALL_KEYS

GRIB_KEYS_ITERATOR_DUMP_ONLY

GRIB_KEYS_ITERATOR_SKIP_CODED

GRIB_KEYS_ITERATOR_SKIP_COMPUTED

GRIB_KEYS_ITERATOR_SKIP_DUPLICATES

GRIB_KEYS_ITERATOR_SKIP_EDITION_SPECIFIC

GRIB_KEYS_ITERATOR_SKIP_FUNCTION

GRIB_KEYS_ITERATOR_SKIP_OPTIONAL

GRIB_KEYS_ITERATOR_SKIP_READ_ONLY

GRIB_LOG_DEBUG

GRIB_LOG_ERROR

GRIB_LOG_FATAL

GRIB_LOG_INFO

GRIB_LOG_WARNING

GRIB_MESSAGE_MALFORMED

GRIB_MESSAGE_TOO_LARGE

GRIB_MISSING_BUFR_ENTRY

GRIB_MISSING_DOUBLE

GRIB_MISSING_KEY

GRIB_MISSING_LONG

GRIB_NEAREST_SAME_DATA

GRIB_NEAREST_SAME_GRID

GRIB_NEAREST_SAME_POINT

GRIB_NOT_FOUND

GRIB_NOT_IMPLEMENTED

GRIB_NO_DEFINITIONS

GRIB_NO_MORE_IN_SET

GRIB_NO_VALUES

GRIB_NULL_HANDLE

GRIB_NULL_INDEX

GRIB_NULL_POINTER

GRIB_OUT_OF_AREA

GRIB_OUT_OF_MEMORY

GRIB_OUT_OF_RANGE

GRIB_PREMATURE_END_OF_FILE

GRIB_READ_ONLY

GRIB_SECTION_BITMAP

GRIB_SECTION_DATA

GRIB_SECTION_GRID

GRIB_SECTION_LOCAL

GRIB_SECTION_PRODUCT

GRIB_STRING_TOO_SMALL

GRIB_SUCCESS

GRIB_SWITCH_NO_MATCH

GRIB_TOO_MANY_ATTRIBUTES

GRIB_TYPE_BYTES

GRIB_TYPE_DOUBLE

GRIB_TYPE_LABEL

GRIB_TYPE_LONG

GRIB_TYPE_MISSING

GRIB_TYPE_SECTION

GRIB_TYPE_STRING

GRIB_TYPE_UNDEFINED

GRIB_UNDERFLOW

GRIB_UNSUPPORTED_EDITION

GRIB_UTIL_ACCURACY_SAME_BITS_PER_VALUES_AS_INPUT

GRIB_UTIL_ACCURACY_SAME_DECIMAL_SCALE_FACTOR_AS_INPUT

GRIB_UTIL_ACCURACY_USE_PROVIDED_BITS_PER_VALUES

GRIB_UTIL_ACCURACY_USE_PROVIDED_DECIMAL_SCALE_FACTOR

GRIB_UTIL_GRID_SPEC_LAMBERT_AZIMUTHAL_EQUAL_AREA

GRIB_UTIL_GRID_SPEC_LAMBERT_CONFORMAL

GRIB_UTIL_GRID_SPEC_POLAR_STEREOGRAPHIC

GRIB_UTIL_GRID_SPEC_REDUCED_GG

GRIB_UTIL_GRID_SPEC_REDUCED_LL

GRIB_UTIL_GRID_SPEC_REDUCED_ROTATED_GG

GRIB_UTIL_GRID_SPEC_REGULAR_GG

GRIB_UTIL_GRID_SPEC_REGULAR_LL

GRIB_UTIL_GRID_SPEC_ROTATED_GG

GRIB_UTIL_GRID_SPEC_ROTATED_LL

GRIB_UTIL_GRID_SPEC_SH

GRIB_UTIL_GRID_SPEC_UNSTRUCTURED

GRIB_UTIL_PACKING_SAME_AS_INPUT

GRIB_UTIL_PACKING_TYPE_CCSDS

GRIB_UTIL_PACKING_TYPE_GRID_COMPLEX

GRIB_UTIL_PACKING_TYPE_GRID_SECOND_ORDER

GRIB_UTIL_PACKING_TYPE_GRID_SIMPLE

GRIB_UTIL_PACKING_TYPE_GRID_SIMPLE_MATRIX

GRIB_UTIL_PACKING_TYPE_JPEG

GRIB_UTIL_PACKING_TYPE_SAME_AS_INPUT

GRIB_UTIL_PACKING_TYPE_SPECTRAL_COMPLEX

GRIB_UTIL_PACKING_TYPE_SPECTRAL_SIMPLE

GRIB_UTIL_PACKING_USE_PROVIDED

GRIB_UTIL_SET_SPEC_FLAGS_ONLY_PACKING

GRIB_VALUE_CANNOT_BE_MISSING

GRIB_VALUE_DIFFERENT

GRIB_WRONG_ARRAY_SIZE

GRIB_WRONG_BITMAP_SIZE

GRIB_WRONG_CONVERSION

GRIB_WRONG_GRID

GRIB_WRONG_LENGTH

GRIB_WRONG_STEP

GRIB_WRONG_STEP_UNIT

GRIB_WRONG_TYPE

LITTLE_ENDIAN

L_ctermid

L_tmpnam

MATH_ERREXCEPT

MATH_ERRNO

M_1_PI

M_2_PI

M_2_SQRTPI

M_E

M_LN2

M_LN10

M_LOG2E

M_LOG10E

M_PI

M_PI_2

M_PI_4

M_SQRT1_2

M_SQRT2

PDP_ENDIAN

P_tmpdir

ProductKind_PRODUCT_ANY

ProductKind_PRODUCT_BUFR

ProductKind_PRODUCT_GRIB

ProductKind_PRODUCT_GTS

ProductKind_PRODUCT_METAR

ProductKind_PRODUCT_TAF

RAND_MAX

SEEK_CUR

SEEK_END

SEEK_SET

TMP_MAX

WCONTINUED

WEXITED

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

_ALLOCA_H

_ATFILE_SOURCE

_BITS_BYTESWAP_H

_BITS_ENDIANNESS_H

_BITS_ENDIAN_H

_BITS_LIBM_SIMD_DECL_STUBS_H

_BITS_PTHREADTYPES_ARCH_H

_BITS_PTHREADTYPES_COMMON_H

_BITS_STDINT_INTN_H

_BITS_STDIO_LIM_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_TYPES_LOCALE_T_H

_BITS_TYPES___LOCALE_T_H

_BITS_UINTN_IDENTITY_H

_DEFAULT_SOURCE

_ENDIAN_H

_FEATURES_H

_IOFBF

_IOLBF

_IONBF

_IO_EOF_SEEN

_IO_ERR_SEEN

_IO_USER_LOCK

_MATH_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDIO_H

_STDLIB_H

_STRINGS_H

_STRING_H

_STRUCT_TIMESPEC

_SYS_CDEFS_H

_SYS_SELECT_H

_SYS_TIME_H

_SYS_TYPES_H

_THREAD_MUTEX_INTERNAL_H

_THREAD_SHARED_TYPES_H

__BIG_ENDIAN

__BIT_TYPES_DEFINED__

__BYTE_ORDER

__ENUM_IDTYPE_T

__FD_SETSIZE

__FD_ZERO_STOS

__FILE_defined

__FLOAT_WORD_ORDER

__FP_LOGB0_IS_MIN

__FP_LOGBNAN_IS_MIN

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNUC_VA_LIST

__GNU_LIBRARY__

__HAVE_DISTINCT_FLOAT16

__HAVE_DISTINCT_FLOAT32

__HAVE_DISTINCT_FLOAT32X

__HAVE_DISTINCT_FLOAT64

__HAVE_DISTINCT_FLOAT64X

__HAVE_DISTINCT_FLOAT128

__HAVE_DISTINCT_FLOAT128X

__HAVE_FLOAT16

__HAVE_FLOAT32

__HAVE_FLOAT32X

__HAVE_FLOAT64

__HAVE_FLOAT64X

__HAVE_FLOAT64X_LONG_DOUBLE

__HAVE_FLOAT128

__HAVE_FLOAT128X

__HAVE_FLOATN_NOT_TYPEDEF

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__LITTLE_ENDIAN

__LONG_DOUBLE_USES_FLOAT128

__MATH_DECLARE_LDOUBLE

__MATH_DECLARING_DOUBLE

__MATH_DECLARING_FLOATN

__OFF_T_MATCHES_OFF64_T

__PDP_ENDIAN

__PTHREAD_MUTEX_HAVE_PREV

__RLIM_T_MATCHES_RLIM64_T

__SIZEOF_PTHREAD_ATTR_T

__SIZEOF_PTHREAD_BARRIERATTR_T

__SIZEOF_PTHREAD_BARRIER_T

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WALL

__WCLONE

__WCOREFLAG

__WNOTHREAD

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__W_CONTINUED

____FILE_defined

_____fpos64_t_defined

_____fpos_t_defined

____mbstate_t_defined

__clock_t_defined

__clockid_t_defined

__glibc_c99_flexarr_available

__have_pthread_attr_t

__itimer_which_ITIMER_PROF

__itimer_which_ITIMER_REAL

__itimer_which_ITIMER_VIRTUAL

__ldiv_t_defined

__lldiv_t_defined

__sigset_t_defined

__struct_FILE_defined

__time_t_defined

__timer_t_defined

__timeval_defined

idtype_t_P_ALL

idtype_t_P_PGID

idtype_t_P_PID

math_errhandling

Statics

signgam

stderr

stdin

stdout

sys_errlist

sys_nerr

Functions

_Exit⚠

__acos⚠

__acosf⚠

__acosh⚠

__acoshf⚠

__acoshl⚠

__acosl⚠

__asin⚠

__asinf⚠

__asinh⚠

__asinhf⚠

__asinhl⚠

__asinl⚠

__atan⚠

__atan2⚠

__atan2f⚠

__atan2l⚠

__atanf⚠

__atanh⚠

__atanhf⚠

__atanhl⚠

__atanl⚠

__cbrt⚠

__cbrtf⚠

__cbrtl⚠

__ceil⚠

__ceilf⚠

__ceill⚠

__copysign⚠

__copysignf⚠

__copysignl⚠

__cos⚠

__cosf⚠

__cosh⚠

__coshf⚠

__coshl⚠

__cosl⚠

__ctype_get_mb_cur_max⚠

__drem⚠

__dremf⚠

__dreml⚠

__erf⚠

__erfc⚠

__erfcf⚠

__erfcl⚠

__erff⚠

__erfl⚠

__exp⚠

__exp2⚠

__exp2f⚠

__exp2l⚠

__expf⚠

__expl⚠

__expm1⚠

__expm1f⚠

__expm1l⚠

__fabs⚠

__fabsf⚠

__fabsl⚠

__fdim⚠

__fdimf⚠

__fdiml⚠

__finite⚠

__finitef⚠

__finitel⚠

__floor⚠

__floorf⚠

__floorl⚠

__fma⚠

__fmaf⚠

__fmal⚠

__fmax⚠

__fmaxf⚠

__fmaxl⚠

__fmin⚠

__fminf⚠

__fminl⚠

__fmod⚠

__fmodf⚠

__fmodl⚠

__fpclassify⚠

__fpclassifyf⚠

__fpclassifyl⚠

__frexp⚠

__frexpf⚠

__frexpl⚠

__gamma⚠

__gammaf⚠

__gammal⚠

__getdelim⚠

__hypot⚠

__hypotf⚠

__hypotl⚠

__ilogb⚠

__ilogbf⚠

__ilogbl⚠

__iseqsig⚠

__iseqsigf⚠

__iseqsigl⚠

__isinf⚠

__isinff⚠

__isinfl⚠

__isnan⚠

__isnanf⚠

__isnanl⚠

__issignaling⚠

__issignalingf⚠

__issignalingl⚠

__j0⚠

__j0f⚠

__j0l⚠

__j1⚠

__j1f⚠

__j1l⚠

__jn⚠

__jnf⚠

__jnl⚠

__ldexp⚠

__ldexpf⚠

__ldexpl⚠

__lgamma⚠

__lgamma_r⚠

__lgammaf⚠

__lgammaf_r⚠

__lgammal⚠

__lgammal_r⚠

__llrint⚠

__llrintf⚠

__llrintl⚠

__llround⚠

__llroundf⚠

__llroundl⚠

__log⚠

__log1p⚠

__log1pf⚠

__log1pl⚠

__log2⚠

__log2f⚠

__log2l⚠

__log10⚠

__log10f⚠

__log10l⚠

__logb⚠

__logbf⚠

__logbl⚠

__logf⚠

__logl⚠

__lrint⚠

__lrintf⚠

__lrintl⚠

__lround⚠

__lroundf⚠

__lroundl⚠

__modf⚠

__modff⚠

__modfl⚠

__nan⚠

__nanf⚠

__nanl⚠

__nearbyint⚠

__nearbyintf⚠

__nearbyintl⚠

__nextafter⚠

__nextafterf⚠

__nextafterl⚠

__nexttoward⚠

__nexttowardf⚠

__nexttowardl⚠

__overflow⚠

__pow⚠

__powf⚠

__powl⚠

__remainder⚠

__remainderf⚠

__remainderl⚠

__remquo⚠

__remquof⚠

__remquol⚠

__rint⚠

__rintf⚠

__rintl⚠

__round⚠

__roundf⚠

__roundl⚠

__scalb⚠

__scalbf⚠

__scalbl⚠

__scalbln⚠

__scalblnf⚠

__scalblnl⚠

__scalbn⚠

__scalbnf⚠

__scalbnl⚠

__signbit⚠

__signbitf⚠

__signbitl⚠

__significand⚠

__significandf⚠

__significandl⚠

__sin⚠

__sinf⚠

__sinh⚠

__sinhf⚠

__sinhl⚠

__sinl⚠

__sqrt⚠

__sqrtf⚠

__sqrtl⚠

__stpcpy⚠

__stpncpy⚠

__strtok_r⚠

__tan⚠

__tanf⚠

__tanh⚠

__tanhf⚠

__tanhl⚠

__tanl⚠

__tgamma⚠

__tgammaf⚠

__tgammal⚠

__trunc⚠

__truncf⚠

__truncl⚠

__uflow⚠

__y0⚠

__y0f⚠

__y0l⚠

__y1⚠

__y1f⚠

__y1l⚠

__yn⚠

__ynf⚠

__ynl⚠

a64l⚠

abort⚠

abs⚠

acos⚠

acosf⚠

acosh⚠

acoshf⚠

acoshl⚠

acosl⚠

adjtime⚠

aligned_alloc⚠

alloca⚠

asin⚠

asinf⚠

asinh⚠

asinhf⚠

asinhl⚠

asinl⚠

at_quick_exit⚠

atan⚠

atan2⚠

atan2f⚠

atan2l⚠

atanf⚠

atanh⚠

atanhf⚠

atanhl⚠

atanl⚠

atexit⚠

atof⚠

atoi⚠

atol⚠

atoll⚠

bcmp⚠

bcopy⚠

bsearch⚠

bzero⚠

calloc⚠

cbrt⚠

cbrtf⚠

cbrtl⚠

ceil⚠

ceilf⚠

ceill⚠

clearenv⚠

clearerr⚠

clearerr_unlocked⚠

codes_box_get_points⚠

codes_box_new⚠

codes_bufr_copy_data⚠

codes_bufr_copy_data_return_copied_keys⚠

codes_bufr_data_section_keys_iterator_new⚠

codes_bufr_extract_headers_malloc⚠

codes_bufr_handle_new_from_file⚠

codes_bufr_handle_new_from_samples⚠

codes_bufr_header_get_string⚠

codes_bufr_keys_iterator_delete⚠

codes_bufr_keys_iterator_get_name⚠

codes_bufr_keys_iterator_new⚠

codes_bufr_keys_iterator_next⚠

codes_bufr_keys_iterator_rewind⚠

codes_bufr_multi_element_constant_arrays_off⚠

codes_bufr_multi_element_constant_arrays_on⚠

codes_check⚠

codes_check_message_footer⚠

codes_check_message_header⚠

codes_context_delete⚠

codes_context_get_default⚠

codes_context_set_definitions_path⚠

codes_context_set_samples_path⚠

codes_copy_key⚠

codes_copy_namespace⚠

codes_count_in_file⚠

codes_count_in_filename⚠

codes_date_to_julian⚠

codes_datetime_to_julian⚠

codes_definition_path⚠

codes_dump_action_tree⚠

codes_dump_content⚠

codes_fieldset_apply_order_by⚠

codes_fieldset_count⚠

codes_fieldset_delete⚠

codes_fieldset_new_from_files⚠

codes_fieldset_next_handle⚠

codes_fieldset_rewind⚠

codes_get_api_version⚠

codes_get_bytes⚠

codes_get_double⚠

codes_get_double_array⚠

codes_get_double_element⚠

codes_get_double_elements⚠

codes_get_error_message⚠

codes_get_gaussian_latitudes⚠

codes_get_git_sha1⚠

codes_get_gribex_mode⚠

codes_get_length⚠

codes_get_long⚠

codes_get_long_array⚠

codes_get_message⚠

codes_get_message_copy⚠

codes_get_message_offset⚠

codes_get_message_size⚠

codes_get_native_type⚠

codes_get_offset⚠

codes_get_package_name⚠

codes_get_product_kind⚠

codes_get_reduced_row⚠

codes_get_reduced_row_p⚠

codes_get_size⚠

codes_get_string⚠

codes_get_string_array⚠

codes_get_type_name⚠

codes_grib_get_data⚠

codes_grib_handle_new_from_file⚠

codes_grib_handle_new_from_multi_message⚠

codes_grib_handle_new_from_samples⚠

codes_grib_iterator_delete⚠

codes_grib_iterator_has_next⚠

codes_grib_iterator_new⚠

codes_grib_iterator_next⚠

codes_grib_iterator_previous⚠

codes_grib_iterator_reset⚠

codes_grib_multi_handle_append⚠

codes_grib_multi_handle_delete⚠

codes_grib_multi_handle_new⚠

codes_grib_multi_handle_write⚠

codes_grib_multi_support_off⚠

codes_grib_multi_support_on⚠

codes_grib_multi_support_reset_file⚠

codes_grib_nearest_delete⚠

codes_grib_nearest_find⚠

codes_grib_nearest_find_multiple⚠

codes_grib_nearest_new⚠

codes_grib_util_get_mars_param⚠

codes_grib_util_get_param_id⚠

codes_grib_util_sections_copy⚠

codes_grib_util_set_spec⚠

codes_gribex_mode_off⚠

codes_gribex_mode_on⚠

codes_gts_header_off⚠

codes_gts_header_on⚠

codes_handle_clone⚠

codes_handle_delete⚠

codes_handle_new_from_file⚠

codes_handle_new_from_index⚠

codes_handle_new_from_message⚠

codes_handle_new_from_message_copy⚠

codes_handle_new_from_partial_message⚠

codes_handle_new_from_partial_message_copy⚠

codes_index_add_file⚠

codes_index_delete⚠

codes_index_get_double⚠

codes_index_get_long⚠

codes_index_get_size⚠

codes_index_get_string⚠

codes_index_new⚠

codes_index_new_from_file⚠

codes_index_read⚠

codes_index_select_double⚠

codes_index_select_long⚠

codes_index_select_string⚠

codes_index_write⚠

codes_is_defined⚠

codes_is_missing⚠

codes_julian_to_date⚠

codes_julian_to_datetime⚠

codes_keys_iterator_delete⚠

codes_keys_iterator_get_bytes⚠

codes_keys_iterator_get_double⚠

codes_keys_iterator_get_long⚠

codes_keys_iterator_get_name⚠

codes_keys_iterator_get_string⚠

codes_keys_iterator_new⚠

codes_keys_iterator_next⚠

codes_keys_iterator_rewind⚠

codes_keys_iterator_set_flags⚠

codes_points_get_values⚠

codes_print_api_version⚠

codes_samples_path⚠

codes_set_bytes⚠

codes_set_codes_assertion_failed_proc⚠

codes_set_double⚠

codes_set_double_array⚠

codes_set_force_double_array⚠

codes_set_long⚠

codes_set_long_array⚠

codes_set_missing⚠

codes_set_string⚠

codes_set_string_array⚠

codes_set_values⚠

codes_update_sections_lengths⚠

codes_values_check⚠

codes_write_message⚠

copysign⚠

copysignf⚠

copysignl⚠

cos⚠

cosf⚠

cosh⚠

coshf⚠

coshl⚠

cosl⚠

ctermid⚠

div⚠

dprintf⚠

drand48⚠

drand48_r⚠

drem⚠

dremf⚠

dreml⚠

ecvt⚠

ecvt_r⚠

erand48⚠

erand48_r⚠

erf⚠

erfc⚠

erfcf⚠

erfcl⚠

erff⚠

erfl⚠

exit⚠

exp⚠

exp2⚠

exp2f⚠

exp2l⚠

expf⚠

expl⚠

explicit_bzero⚠

expm1⚠

expm1f⚠

expm1l⚠

fabs⚠

fabsf⚠

fabsl⚠

fclose⚠

fcvt⚠

fcvt_r⚠

fdim⚠

fdimf⚠

fdiml⚠

fdopen⚠

feof⚠

feof_unlocked⚠

ferror⚠

ferror_unlocked⚠

fflush⚠

fflush_unlocked⚠

ffs⚠

ffsl⚠

ffsll⚠

fgetc⚠

fgetc_unlocked⚠

fgetpos⚠

fgets⚠

fileno⚠

fileno_unlocked⚠

finite⚠

finitef⚠

finitel⚠

flockfile⚠

floor⚠

floorf⚠

floorl⚠

fma⚠

fmaf⚠

fmal⚠

fmax⚠

fmaxf⚠

fmaxl⚠

fmemopen⚠

fmin⚠

fminf⚠

fminl⚠

fmod⚠

fmodf⚠

fmodl⚠

fopen⚠

fprintf⚠

fputc⚠

fputc_unlocked⚠

fputs⚠

fread⚠

fread_unlocked⚠

free⚠

freopen⚠

frexp⚠

frexpf⚠

frexpl⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

ftell⚠

ftello⚠

ftrylockfile⚠

funlockfile⚠

futimes⚠

fwrite⚠

fwrite_unlocked⚠

gamma⚠

gammaf⚠

gammal⚠

gcvt⚠

getc⚠

getc_unlocked⚠

getchar⚠

getchar_unlocked⚠

getdelim⚠

getenv⚠

getitimer⚠

getline⚠

getloadavg⚠

getsubopt⚠

gettimeofday⚠

getw⚠

grib_box_get_points⚠

grib_box_new⚠

grib_check⚠

grib_context_delete⚠

grib_context_get_default⚠

grib_context_set_buffer_memory_proc⚠

grib_context_set_definitions_path⚠

grib_context_set_logging_proc⚠

grib_context_set_memory_proc⚠

grib_context_set_persistent_memory_proc⚠

grib_context_set_print_proc⚠

grib_context_set_samples_path⚠

grib_copy_namespace⚠

grib_count_in_file⚠

grib_count_in_filename⚠

grib_date_to_julian⚠

grib_datetime_to_julian⚠

grib_definition_path⚠

grib_dump_action_tree⚠

grib_dump_content⚠

grib_fieldset_apply_order_by⚠

grib_fieldset_count⚠

grib_fieldset_delete⚠

grib_fieldset_new_from_files⚠

grib_fieldset_next_handle⚠

grib_fieldset_rewind⚠

grib_get_api_version⚠

grib_get_bytes⚠

grib_get_data⚠

grib_get_double⚠

grib_get_double_array⚠

grib_get_double_element⚠

grib_get_double_elements⚠

grib_get_error_message⚠

grib_get_gaussian_latitudes⚠

grib_get_git_sha1⚠

grib_get_gribex_mode⚠

grib_get_length⚠

grib_get_long⚠

grib_get_long_array⚠

grib_get_message⚠

grib_get_message_copy⚠

grib_get_message_offset⚠

grib_get_message_size⚠

grib_get_native_type⚠

grib_get_offset⚠

grib_get_package_name⚠

grib_get_reduced_row⚠

grib_get_reduced_row_p⚠

grib_get_size⚠

grib_get_string⚠

grib_get_string_array⚠

grib_get_type_name⚠

grib_gribex_mode_off⚠

grib_gribex_mode_on⚠

grib_gts_header_off⚠

grib_gts_header_on⚠

grib_handle_clone⚠

grib_handle_delete⚠

grib_handle_new_from_file⚠

grib_handle_new_from_index⚠

grib_handle_new_from_message⚠

grib_handle_new_from_message_copy⚠

grib_handle_new_from_multi_message⚠

grib_handle_new_from_partial_message⚠

grib_handle_new_from_partial_message_copy⚠

grib_handle_new_from_samples⚠

grib_index_add_file⚠

grib_index_delete⚠

grib_index_get_double⚠

grib_index_get_long⚠

grib_index_get_size⚠

grib_index_get_string⚠

grib_index_new⚠

grib_index_new_from_file⚠

grib_index_read⚠

grib_index_select_double⚠

grib_index_select_long⚠

grib_index_select_string⚠

grib_index_write⚠

grib_is_defined⚠

grib_is_missing⚠

grib_iterator_delete⚠

grib_iterator_has_next⚠

grib_iterator_new⚠

grib_iterator_next⚠

grib_iterator_previous⚠

grib_iterator_reset⚠

grib_julian_to_date⚠

grib_julian_to_datetime⚠

grib_keys_iterator_delete⚠

grib_keys_iterator_get_bytes⚠

grib_keys_iterator_get_double⚠

grib_keys_iterator_get_long⚠

grib_keys_iterator_get_name⚠

grib_keys_iterator_get_string⚠

grib_keys_iterator_new⚠

grib_keys_iterator_next⚠

grib_keys_iterator_rewind⚠

grib_keys_iterator_set_flags⚠

grib_multi_handle_append⚠

grib_multi_handle_delete⚠

grib_multi_handle_new⚠

grib_multi_handle_write⚠

grib_multi_support_off⚠

grib_multi_support_on⚠

grib_multi_support_reset_file⚠

grib_nearest_delete⚠

grib_nearest_find⚠

grib_nearest_find_multiple⚠

grib_nearest_new⚠

grib_new_from_file⚠

grib_points_get_values⚠

grib_print_api_version⚠

grib_read_any_from_file⚠

grib_samples_path⚠

grib_set_bytes⚠

grib_set_double⚠

grib_set_double_array⚠

grib_set_force_double_array⚠

grib_set_long⚠

grib_set_long_array⚠

grib_set_missing⚠

grib_set_string⚠

grib_set_string_array⚠

grib_set_values⚠

grib_update_sections_lengths⚠

grib_util_get_mars_param⚠

grib_util_get_param_id⚠

grib_util_sections_copy⚠

grib_util_set_spec⚠

grib_util_set_spec2⚠

grib_values_check⚠

grib_write_message⚠

hypot⚠

hypotf⚠

hypotl⚠

ilogb⚠

ilogbf⚠

ilogbl⚠

index⚠

initstate⚠

initstate_r⚠

isinf⚠

isinff⚠

isinfl⚠

isnan⚠

isnanf⚠

isnanl⚠

j0⚠

j0f⚠

j0l⚠

j1⚠

j1f⚠

j1l⚠

jn⚠

jnf⚠

jnl⚠

jrand48⚠

jrand48_r⚠

l64a⚠

labs⚠

lcong48⚠

lcong48_r⚠

ldexp⚠

ldexpf⚠

ldexpl⚠

ldiv⚠

lgamma⚠

lgamma_r⚠

lgammaf⚠

lgammaf_r⚠

lgammal⚠

lgammal_r⚠

llabs⚠

lldiv⚠

llrint⚠

llrintf⚠

llrintl⚠

llround⚠

llroundf⚠

llroundl⚠

log⚠

log1p⚠

log1pf⚠

log1pl⚠

log2⚠

log2f⚠

log2l⚠

log10⚠

log10f⚠

log10l⚠

logb⚠

logbf⚠

logbl⚠

logf⚠

logl⚠

lrand48⚠

lrand48_r⚠

lrint⚠

lrintf⚠

lrintl⚠

lround⚠

lroundf⚠

lroundl⚠

lutimes⚠

malloc⚠

mblen⚠

mbstowcs⚠

mbtowc⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mkdtemp⚠

mkstemp⚠

mkstemps⚠

mktemp⚠

modf⚠

modff⚠

modfl⚠

mrand48⚠

mrand48_r⚠

nan⚠

nanf⚠

nanl⚠

nearbyint⚠

nearbyintf⚠

nearbyintl⚠

nextafter⚠

nextafterf⚠

nextafterl⚠

nexttoward⚠

nexttowardf⚠

nexttowardl⚠

nrand48⚠

nrand48_r⚠

on_exit⚠

open_memstream⚠

parse_keyval_string⚠

pclose⚠

perror⚠

popen⚠

posix_memalign⚠

pow⚠

powf⚠

powl⚠

printf⚠

pselect⚠

putc⚠

putc_unlocked⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

putw⚠

qecvt⚠

qecvt_r⚠

qfcvt⚠

qfcvt_r⚠

qgcvt⚠

qsort⚠

quick_exit⚠

rand⚠

rand_r⚠

random⚠

random_r⚠

realloc⚠

reallocarray⚠

realpath⚠

remainder⚠

remainderf⚠

remainderl⚠

remove⚠

remquo⚠

remquof⚠

remquol⚠

rename⚠

renameat⚠

rewind⚠

rindex⚠

rint⚠

rintf⚠

rintl⚠

round⚠

roundf⚠

roundl⚠

rpmatch⚠

scalb⚠

scalbf⚠

scalbl⚠

scalbln⚠

scalblnf⚠

scalblnl⚠

scalbn⚠

scalbnf⚠

scalbnl⚠

scanf⚠

seed48⚠

seed48_r⚠

select⚠

setbuf⚠

setbuffer⚠

setenv⚠

setitimer⚠

setlinebuf⚠

setstate⚠

setstate_r⚠

settimeofday⚠

setvbuf⚠

significand⚠

significandf⚠

significandl⚠

sin⚠

sinf⚠

sinh⚠

sinhf⚠

sinhl⚠

sinl⚠

snprintf⚠

sprintf⚠

sqrt⚠

sqrtf⚠

sqrtl⚠

srand⚠

srand48⚠

srand48_r⚠

srandom⚠

srandom_r⚠

sscanf⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasecmp_l⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcoll_l⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_l⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncasecmp_l⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsep⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtold⚠

strtoll⚠

strtoq⚠

strtoul⚠

strtoull⚠

strtouq⚠

strxfrm⚠

strxfrm_l⚠

system⚠

tan⚠

tanf⚠

tanh⚠

tanhf⚠

tanhl⚠

tanl⚠

tempnam⚠

tgamma⚠

tgammaf⚠

tgammal⚠

tmpfile⚠

tmpnam⚠

tmpnam_r⚠

trunc⚠

truncf⚠

truncl⚠

ungetc⚠

unsetenv⚠

utimes⚠

valloc⚠

vdprintf⚠

vfprintf⚠

vfscanf⚠

vprintf⚠

vscanf⚠

vsnprintf⚠

vsprintf⚠

vsscanf⚠

wcstombs⚠

wctomb⚠

wmo_read_any_from_file⚠

wmo_read_any_from_file_malloc⚠

wmo_read_any_from_stream⚠

wmo_read_any_from_stream_malloc⚠

wmo_read_bufr_from_file⚠

wmo_read_bufr_from_file_malloc⚠

wmo_read_grib_from_file⚠

wmo_read_grib_from_file_malloc⚠

wmo_read_gts_from_file⚠

wmo_read_gts_from_file_malloc⚠

y0⚠

y0f⚠

y0l⚠

y1⚠

y1f⚠

y1l⚠

yn⚠

ynf⚠

ynl⚠

Type Definitions

FILE

ProductKind

_Float32

_Float32x

_Float64

_Float64x

_IO_lock_t

__FILE

__blkcnt64_t

__blkcnt_t

__blksize_t

__builtin_va_list

__caddr_t

__clock_t

__clockid_t

__compar_fn_t

__daddr_t

__dev_t

__fd_mask

__fpos64_t

__fpos_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__gnuc_va_list

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__itimer_which

__itimer_which_t

__key_t

__locale_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__pthread_list_t

__pthread_slist_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

_bindgen_ty_1

blkcnt_t

blksize_t

caddr_t

clock_t

clockid_t

codes_assertion_failed_proc

codes_box

codes_bufr_keys_iterator

codes_context

codes_darray

codes_fieldset

codes_handle

codes_iarray

codes_index

codes_iterator

codes_key_value_list

codes_keys_iterator

codes_multi_handle

codes_nearest

codes_oarray

codes_order_by

codes_points

codes_sarray

codes_string_list

codes_util_grid_spec

codes_util_packing_spec

codes_values

codes_vdarray

codes_viarray

codes_vsarray

codes_where

daddr_t

dev_t

double_t

fd_mask

float_t

fpos_t

fsblkcnt_t

fsfilcnt_t

fsid_t

gid_t

grib_data_eof_proc

grib_data_read_proc

grib_data_seek_proc

grib_data_tell_proc

grib_data_write_proc

grib_free_proc

grib_log_proc

grib_malloc_proc

grib_print_proc

grib_realloc_proc

id_t

idtype_t

ino_t

key_t

locale_t

loff_t

mode_t

nlink_t

off_t

pid_t

pthread_key_t

pthread_once_t

pthread_spinlock_t

pthread_t

quad_t

register_t

sigset_t

size_t

ssize_t

suseconds_t

time_t

timer_t

u_char

u_int

u_int8_t

u_int16_t

u_int32_t

u_int64_t

u_long

u_quad_t

u_short

uid_t

uint

ulong

ushort

va_list

wchar_t

Unions

__mbstate_t__bindgen_ty_1

__pthread_cond_s__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2

pthread_attr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t